ast: split context into separate file
Each of these files use their own forward declarations, and individual
implementations need only include the file that matters. This paves the
way for some additional functionality in the ASTContext itself, and
separates concerns for individual AST nodes.

This is necessary because we will need to deal with multiple soruce
files in the future to support `import`. Currently, the top-level file
calls `set_source` on the log singleton. However, this will not scale as
additional files are parsed: my intent is to hang the raw source off
each `ASTContext` and/or the future `Diagnostics` class.
