+++ src/ast/ast.cpp
@@ -2,6 +2,7 @@
 
 #include <algorithm>
 
+#include "ast/context.h"
 #include "ast/visitor.h"
 #include "log.h"
