+++ src/ast/ast.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <cstdint>
-#include <map>
 #include <string>
 #include <vector>
 
@@ -511,36 +510,6 @@ std::string opstr(const Jump &jump);
 
 SizedType ident_to_record(const std::string &ident, int pointer_level = 0);
 SizedType ident_to_sized_type(const std::string &ident);
-
-template <typename T>
-concept NodeType = std::derived_from<T, Node>;
-
-// Manages the lifetime of AST nodes.
-//
-// Nodes allocated by an ASTContext will be kept alive for the duration of the
-// owning ASTContext object.
-class ASTContext {
-public:
-  Program *root = nullptr;
-
-  // Creates and returns a pointer to an AST node.
-  template <NodeType T, typename... Args>
-  T *make_node(Args &&...args)
-  {
-    auto uniq_ptr = std::make_unique<T>(std::forward<Args>(args)...);
-    auto *raw_ptr = uniq_ptr.get();
-    nodes_.push_back(std::move(uniq_ptr));
-    return raw_ptr;
-  }
-
-  unsigned int node_count()
-  {
-    return nodes_.size();
-  }
-
-private:
-  std::vector<std::unique_ptr<Node>> nodes_;
-};
 
 } // namespace ast
 } // namespace bpftrace
